# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Voetbalclub Kantine Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List all shifts
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: seasonId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftWithAssignments"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Create a shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShiftRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftWithAssignments"
  /shifts/{id}:
    get:
      operationId: getShift
      tags: [shifts]
      summary: Get a shift by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftWithAssignments"
        "404":
          description: Not found
    put:
      operationId: updateShift
      tags: [shifts]
      summary: Update a shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShiftRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftWithAssignments"
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete a shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /shifts/{id}/assign:
    post:
      operationId: assignVolunteer
      tags: [assignments]
      summary: Assign a volunteer to a shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignVolunteerRequest"
      responses:
        "201":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftWithAssignments"
  /shifts/{id}/unassign/{volunteerId}:
    delete:
      operationId: unassignVolunteer
      tags: [assignments]
      summary: Remove a volunteer from a shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: volunteerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unassigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftWithAssignments"
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    Volunteer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        hasPassword:
          type: boolean
          description: "True als er een wachtwoord is ingesteld voor dit account"
        isAdmin:
          type: boolean
          description: "True als deze vrijwilliger beheerdersrechten heeft"
        availability:
          type: array
          items:
            $ref: "#/components/schemas/ShiftSlot"
        groupId:
          type: integer
          nullable: true
        groupMembers:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
            required: [id, name]
        createdAt:
          type: string
      required: [id, name, availability, groupMembers, createdAt]
...
    ShiftSlot:
      type: string
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        seasonId:
          type: integer
          nullable: true
        date:
          type: string
        slot:
          $ref: "#/components/schemas/ShiftSlot"
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        capacity:
          type: integer
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, date, slot, capacity, createdAt]
    ShiftWithAssignments:
      allOf:
        - $ref: "#/components/schemas/Shift"
        - type: object
          properties:
            assignments:
              type: array
              items:
                $ref: "#/components/schemas/Assignment"
          required: [assignments]
    Assignment:
      type: object
      properties:
        id:
          type: integer
        shiftId:
          type: integer
        volunteerId:
          type: integer
        volunteer:
          $ref: "#/components/schemas/Volunteer"
        createdAt:
          type: string
      required: [id, shiftId, volunteerId, volunteer, createdAt]
    CreateShiftRequest:
      type: object
      properties:
        seasonId:
          type: integer
          nullable: true
        date:
          type: string
        slot:
          $ref: "#/components/schemas/ShiftSlot"
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        capacity:
          type: integer
        notes:
          type: string
          nullable: true
      required: [date, slot, capacity]
    UpdateShiftRequest:
      type: object
      properties:
        date:
          type: string
        slot:
          $ref: "#/components/schemas/ShiftSlot"
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        capacity:
          type: integer
        notes:
          type: string
          nullable: true
...
    AssignVolunteerRequest:
      type: object
      properties:
        volunteerId:
          type: integer
      required: [volunteerId]
...